--- zojae031/Project/app/src/main/java/study/architecture/presentation/CoinAdapterContract.kt
@@ -1,15 +1,0 @@
-package study.architecture.presentation
-
-import study.architecture.data.entity.ProcessingTicker
-
-interface CoinAdapterContract {
-    interface View {
-        fun updateList(list: List<ProcessingTicker>)
-        fun clearList()
-    }
-
-    interface Model {
-        fun notifyDataChange()
-    }
-
-}
