--- zojae031/Project/app/src/main/java/study/architecture/presentation/CoinContract.kt
@@ -1,16 +1,0 @@
-package study.architecture.presentation
-
-interface CoinContract {
-    interface View {
-        fun showProgress()
-        fun hideProgress()
-        fun showError(e: String?)
-    }
-
-    interface Presenter {
-        fun onResume()
-        fun onPause()
-        fun setAdapterView(adapterView: CoinAdapterContract.View)
-        fun setAdapterModel(adapterModel: CoinAdapterContract.Model)
-    }
-}
