+++ zojae031/Project/app/src/main/java/study/architecture/ui/coinjob/CoinViewModel.kt
@@ -1,32 +1,22 @@
-package study.architecture.presentation
+package study.architecture.ui.coinjob
 
-import android.annotation.SuppressLint
+import android.util.Log
 import io.reactivex.android.schedulers.AndroidSchedulers
 import io.reactivex.disposables.CompositeDisposable
 import io.reactivex.disposables.Disposable
 import study.architecture.data.entity.ProcessingTicker
 import study.architecture.data.repository.RepositoryImpl
-import study.architecture.ui.coinjob.CoinFragment
 import study.architecture.util.TextUtil
 import java.util.concurrent.TimeUnit
 
-/**
- * 1. 업비트 데이터를 가져와 View에게 알려준다.
- * 2. Observable 생명주기를 관리한다.
- */
-class CoinPresenter(
-    private val view: CoinContract.View,
+class CoinViewModel(
     private val index: CoinFragment.FragIndex,
-    private val repository: RepositoryImpl
-) :
-    CoinContract.Presenter {
+    private val repository: RepositoryImpl,
+    private val adapter: CoinDataAdapter
+) {
     private lateinit var marketName: String
-
-    private val dispose: Disposable
     private val compositeDisposable = CompositeDisposable()
-
-    private lateinit var adapterView: CoinAdapterContract.View
-    private lateinit var adapterModel: CoinAdapterContract.Model
+    private val dispose: Disposable
 
     init {
         repository.getMarkets()
@@ -46,31 +36,10 @@ class CoinPresenter(
                     tickerRequest()
                 },
                 {
-                    view.showError(it.message)
+                    Log.e("CoinViewModelError", it.message)
                 }).also { dispose = it }
     }
 
-    override fun setAdapterView(adapterView: CoinAdapterContract.View) {
-        this.adapterView = adapterView
-    }
-
-    override fun setAdapterModel(adapterModel: CoinAdapterContract.Model) {
-        this.adapterModel = adapterModel
-    }
-
-    override fun onResume() {
-        if (dispose.isDisposed) {
-            tickerRequest()
-        }
-    }
-
-
-    override fun onPause() {
-        dispose.dispose()
-        compositeDisposable.clear()
-    }
-
-    @SuppressLint("CheckResult")
     private fun tickerRequest() {
         repository.getTickers(marketName)
             .repeatWhen { it.delay(8, TimeUnit.SECONDS) }
@@ -91,9 +60,10 @@ class CoinPresenter(
             .subscribe(
                 { lists ->
                     if (!repository.checkNetwork()) view.showError("데이터가 최신이 아닙니다.\n 인터넷에 연결해주세요")
-                    adapterView.clearList()
-                    adapterView.updateList(lists)
-                    adapterModel.notifyDataChange()
+
+                    adapter.clearList()
+                    adapter.updateList(lists)
+                    adapter.notifyDataChange()
                 },
                 { e ->
                     view.showError(e.message)
@@ -101,6 +71,4 @@ class CoinPresenter(
             ).also { compositeDisposable.add(it) }
 
     }
-
-
 }
